# app/govern/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bll.AddMenuReq:
    properties:
      enable:
        type: boolean
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      parentID:
        type: string
      route:
        type: string
      seq:
        type: integer
      show:
        type: boolean
      widgets:
        items:
          $ref: '#/definitions/bll.AddWidget'
        type: array
    required:
    - name
    - route
    - seq
    type: object
...
  bll.AddWidget:
    properties:
      api:
        type: string
      enable:
        type: boolean
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      seq:
        type: integer
      show:
        type: boolean
    required:
    - api
    - name
    - seq
    type: object
  bll.AddWidgetReq:
    properties:
      api:
        type: string
      enable:
        type: boolean
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      parentID:
        type: string
      seq:
        type: integer
      show:
        type: boolean
    required:
    - api
    - name
    - seq
    type: object
...
  bll.EnableMenuReq:
    properties:
      enable:
        type: boolean
    required:
    - enable
    type: object
...
  bll.EnableWidgetReq:
    properties:
      enable:
        type: boolean
    required:
    - enable
    type: object
...
  bll.ListMenuResp:
    properties:
      data:
        type: object
      total:
        type: integer
    type: object
...
  bll.ListWidgetResp:
    properties:
      data:
        type: object
      total:
        type: integer
    type: object
...
  bll.ProfileMenuResp:
    properties:
      createdAt:
        type: integer
      creator:
        type: string
      enable:
        type: boolean
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      parentID:
        type: string
      route:
        type: string
      seq:
        type: integer
      show:
        type: boolean
      updatedAt:
        type: integer
    type: object
...
  bll.ProfileWidgetResp:
    properties:
      api:
        properties:
          id:
            type: string
          method:
            type: string
          path:
            type: string
        type: object
      createdAt:
        type: integer
      creator:
        type: string
      enable:
        type: boolean
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      seq:
        type: integer
      show:
        type: boolean
      updatedAt:
        type: integer
    type: object
...
  bll.UpdateMenuReq:
    properties:
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      parentID:
        type: string
      route:
        type: string
      seq:
        type: integer
      show:
        type: boolean
    type: object
...
  bll.UpdateWidgetReq:
    properties:
      api:
        type: string
      icon:
        type: string
      memo:
        type: string
      name:
        type: string
      seq:
        type: integer
      show:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: sfshf@github.com
    name: gavin
    url: http://github.com/sfshf
  description: This is a back-end government app.
  license:
    name: MIT
    url: https://github.com/sfshf/sprout/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Govern APIs
  version: 0.0.1-beta
paths:
...
  /menus:
    get:
      description: Get a list of menu.
      operationId: menu-list
      parameters:
      - in: query
        name: enable
        type: boolean
      - in: query
        name: name
        type: string
      - in: query
        name: route
        type: string
      - in: query
        name: show
        type: boolean
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: menu list.
          schema:
            $ref: '#/definitions/bll.ListMenuResp'
      security:
      - ApiKeyAuth: []
      summary: Get a list of menu.
      tags:
      - menu
    post:
      consumes:
      - application/json
      description: Add a new menu, with widgets if has.
      operationId: menu-add
      parameters:
      - description: required attributes to add a new menu.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.AddMenuReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Add a new menu, with widgets if has.
      tags:
      - menu
  /menus/:id:
    delete:
      description: Evict a specific menu.
      operationId: menu-evict
      parameters:
      - description: id of the menu to evict.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Evict a specific menu.
      tags:
      - menu
    get:
      description: Get the profile of a menu.
      operationId: menu-profile
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: profile of the api.
          schema:
            $ref: '#/definitions/bll.ProfileMenuResp'
      security:
      - ApiKeyAuth: []
      summary: Get infos of a menu.
      tags:
      - menu
    put:
      consumes:
      - application/json
      description: Update a specific menu.
      operationId: menu-update
      parameters:
      - description: id of the menu to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateMenuReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update a specific menu.
      tags:
      - menu
  /menus/:id/enable:
    patch:
      consumes:
      - application/json
      description: Enable or disable a menu.
      operationId: menu-enable
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      - description: true for enable, or false for disable.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.EnableMenuReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Enable or disable a menu.
      tags:
      - menu
  /menus/:id/widgets:
    get:
      description: Get a widget list of a specific menu.
      operationId: menu-profile-widgetList
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: widget list of a specific menu.
          schema:
            $ref: '#/definitions/bll.ListWidgetResp'
      security:
      - ApiKeyAuth: []
      summary: Get a widget list of a specific menu.
      tags:
      - menu
    post:
      consumes:
      - application/json
      description: Add a widget for a specific menu.
      operationId: menu-update-widget-add
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      - description: necessary attributes to add a widget.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.AddWidgetReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Add a widget for a specific menu.
      tags:
      - menu
  /menus/:id/widgets/:widgetId:
    delete:
      consumes:
      - application/json
      description: Evict a widget for a specific menu.
      operationId: menu-update-widget-evict
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      - description: id of the widget.
        in: path
        name: widgetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Evict a widget for a specific menu.
      tags:
      - menu
    get:
      description: Get the profile of a widget.
      operationId: menu-profile-widget-profile
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      - description: id the a widget.
        in: path
        name: widgetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: profile of the widget.
          schema:
            $ref: '#/definitions/bll.ProfileWidgetResp'
      security:
      - ApiKeyAuth: []
      summary: Get infos of a widget.
      tags:
      - menu
    put:
      consumes:
      - application/json
      description: Update infos of a widget.
      operationId: menu-update-widget-update
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      - description: id the a widget.
        in: path
        name: widgetId
        required: true
        type: string
      - description: some attributes to update.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateWidgetReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update infos of a widget.
      tags:
      - menu
  /menus/:id/widgets/:widgetId/enable:
    patch:
      description: Enable or disable a widget.
      operationId: menu-update-widget-profile
      parameters:
      - description: id of the menu.
        in: path
        name: id
        required: true
        type: string
      - description: id the a widget.
        in: path
        name: widgetId
        required: true
        type: string
      - description: true for enable, or false for disable.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.EnableWidgetReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Enable or disable a widget.
      tags:
      - menu
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"